[MOVE/#1] 패키지 정리 및 xml layout 태그 추가
