+++ app/src/main/java/org/android/go/sopt/presentation/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package org.android.go.sopt
+package org.android.go.sopt.presentation.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import org.android.go.sopt.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
