+++ app/src/main/res/layout/activity_login.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".presentation.login.LoginActivity">
 
     <data>
 
@@ -11,8 +12,7 @@
         android:id="@+id/layout_login"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:paddingHorizontal="40dp"
-        tools:context=".LoginActivity">
+        android:paddingHorizontal="40dp">
 
         <TextView
             android:id="@+id/tv_login_welcome_msg"
